+++ src/data/import_manifest.py
@@ -9,15 +9,10 @@ Created on Tue Mar  2 10:38:05 2021
 import click
 import logging
 
-from typing import Union
-
-from mongoengine.errors import NotUniqueError
-from mongoengine.queryset import QuerySet
-
-from src.features.illumina import read_snpChip
+from src.features.illumina import read_Manifest
 from src.features.smarterdb import (
-    VariantSheep, Location, global_connection, IlluminaChip, VariantGoat)
-from src.data.common import get_variant_species
+    Location, global_connection, SupportedChip)
+from src.data.common import get_variant_species, update_variant, new_variant
 
 logger = logging.getLogger(__name__)
 
@@ -33,7 +28,7 @@ def main(species, manifest, chip_name, version, sender):
     VariantSpecie = get_variant_species(species)
 
     # check chip_name
-    illumina_chip = IlluminaChip.objects(name=chip_name).get()
+    illumina_chip = SupportedChip.objects(name=chip_name).get()
 
     # reset chip data (if any)
     illumina_chip.n_of_snps = 0
@@ -41,30 +36,31 @@ def main(species, manifest, chip_name, version, sender):
     logger.info(f"Reading from {manifest}")
 
     # grep a sample SNP
-    for i, snpchip in enumerate(read_snpChip(manifest, delimiter=",")):
+    for i, record in enumerate(read_Manifest(manifest, delimiter=",")):
         # update chip data indipendentely if it is an update or not
         illumina_chip.n_of_snps += 1
 
         # create a location object
         location = Location(
             version=version,
-            chrom=snpchip.chr,
-            position=snpchip.mapinfo,
-            illumina=snpchip.snp,
-            illumina_strand=snpchip.ilmnstrand,
-            strand=snpchip.sourcestrand,
-            imported_from="manifest"
+            chrom=record.chr,
+            position=record.mapinfo,
+            illumina=record.snp,
+            illumina_strand=record.ilmnstrand,
+            strand=record.sourcestrand,
+            imported_from="manifest",
+            date=record.date,
         )
 
         variant = VariantSpecie(
             chip_name=[chip_name],
-            name=snpchip.name,
-            sequence=snpchip.sourceseq,
+            name=record.name,
+            sequence={'manifest': record.sourceseq},
             sender=sender
         )
 
         # search for a snp in database (relying on name)
-        qs = VariantSpecie.objects.filter(name=snpchip.name)
+        qs = VariantSpecie.objects.filter(name=record.name)
 
         if qs.count() == 1:
             update_variant(qs, variant, location)
@@ -81,69 +77,6 @@ def main(species, manifest, chip_name, version, sender):
     logger.info(f"{i+1} variants processed")
 
     logger.info("Completed")
-
-
-def update_variant(
-        qs: QuerySet,
-        variant: Union[VariantSheep, VariantGoat],
-        location: Location):
-    """Update an existing variant (if necessary)"""
-
-    record = qs.get()
-    logger.debug(f"found {record} in database")
-
-    # check chip_name in variant list
-    record = update_chip_name(variant, record)
-
-    # I chose to not update other values, I suppose they be the same
-    # However check for locations
-    check_location(location, record)
-
-
-def update_chip_name(variant, record):
-    variant_set = set(variant.chip_name)
-    record_set = set(record.chip_name)
-
-    # get new items as a difference of two sets
-    new_chips = variant_set - record_set
-
-    if len(new_chips) > 0:
-        # this will append the resulting set as a list
-        record.chip_name += list(new_chips)
-        record.save()
-
-    return record
-
-
-def check_location(location, variant):
-    # get the old location as index
-    index = variant.get_location_index(
-        version=location.version, imported_from=location.imported_from)
-
-    # ok get the old location and check with the new one
-    if variant.locations[index] == location:
-        logger.debug("Locations match")
-
-    # HINT: should I update location?
-    else:
-        logger.warning(
-            f"Locations differ: {location} <> {variant.locations[index]}")
-
-
-def new_variant(
-        variant: Union[VariantSheep, VariantGoat],
-        location: Location):
-
-    variant.locations.append(location)
-
-    logger.info(f"adding {variant} to database")
-
-    try:
-        variant.save()
-
-    except NotUniqueError as e:
-        logger.error(
-            f"Cannot insert {variant}, reason: {e}")
 
 
 if __name__ == '__main__':