+++ tests/data/test_common.py
@@ -16,11 +16,12 @@ from openpyxl import Workbook
 
 from src.data.common import (
     fetch_and_check_dataset, get_variant_species, get_sample_species,
-    pandas_open)
+    pandas_open, update_chip_name, update_sequence, update_affymetrix_record,
+    update_location, update_variant, update_rs_id)
 from src.features.smarterdb import (
-    Dataset, VariantGoat, VariantSheep, SampleSheep, SampleGoat)
+    Dataset, VariantGoat, VariantSheep, SampleSheep, SampleGoat, Location)
 
-from ..common import MongoMockMixin, SmarterIDMixin
+from ..common import MongoMockMixin, SmarterIDMixin, VariantsMixin
 
 FEATURE_DATA_DIR = pathlib.Path(__file__).parents[1] / "features/data"
 SCRIPTS_DATA_DIR = pathlib.Path(__file__).parents[1] / "data/data"
@@ -203,6 +204,146 @@ class PandasOpenTest(unittest.TestCase):
 
             data = pandas_open(datapath)
             self.assertIsInstance(data, pd.DataFrame)
+
+
+class VariantUpdateTests(VariantsMixin, MongoMockMixin, unittest.TestCase):
+    def setUp(self):
+        self.record = VariantSheep.objects.get(
+            name="250506CS3900065000002_1238.1")
+
+        # make a copy, those are all refrences
+        variant_data = self.data[0].copy()
+        location_data = variant_data.pop('locations')[0]
+
+        self.variant = VariantSheep(**variant_data)
+        self.location = Location(**location_data)
+
+    def tearDown(self):
+        # reset record to its original state
+        self.record.update(**self.data[0])
+
+    def test_update_chip_name(self):
+        record, updated = update_chip_name(self.variant, self.record)
+        self.assertFalse(updated)
+
+        # set a new chip_name
+        self.variant.chip_name = ["test"]
+        record, updated = update_chip_name(self.variant, self.record)
+        self.assertTrue(updated)
+
+    def test_update_sequence(self):
+        record, updated = update_sequence(self.variant, self.record)
+        self.assertFalse(updated)
+
+        # delete sequence from variant, no update
+        self.variant.sequence = None
+        record, updated = update_sequence(self.variant, self.record)
+        self.assertFalse(updated)
+
+        # update sequence
+        self.variant.sequence = {"test": "AGCT"}
+        record, updated = update_sequence(self.variant, self.record)
+        self.assertTrue(updated)
+
+    def test_update_affymetrix_record(self):
+        record, updated = update_affymetrix_record(self.variant, self.record)
+        self.assertFalse(updated)
+
+        # add affymetrix attribute
+        self.variant.probeset_id = "test"
+        self.variant.affy_snp_id = "test"
+
+        record, updated = update_affymetrix_record(self.variant, self.record)
+        self.assertTrue(updated)
+
+    def test_update_location(self):
+        record, updated = update_location(self.location, self.record)
+        self.assertFalse(updated)
+
+        # change location
+        self.location.version = "test"
+        record, updated = update_location(self.location, self.record)
+        self.assertTrue(updated)
+
+    def test_update_location_mismatch(self):
+        # test a location mismatch
+        # HINT: should I overwrite the location?
+        self.location.chrom = "test"
+        record, updated = update_location(self.location, self.record)
+        self.assertFalse(updated)
+
+        # no location added
+        self.assertEqual(len(self.record.locations), 2)
+
+    def test_update_rs_id(self):
+        record, updated = update_rs_id(self.variant, self.record)
+        self.assertFalse(updated)
+
+        # unset rs_id
+        self.variant.rs_id = None
+        record, updated = update_rs_id(self.variant, self.record)
+        self.assertFalse(updated)
+
+        # udpate rs_id
+        self.variant.rs_id = "test"
+        record, updated = update_rs_id(self.variant, self.record)
+        self.assertTrue(updated)
+
+    def test_update_variant_chip_name(self):
+        qs = VariantSheep.objects.filter(
+            name="250506CS3900065000002_1238.1")
+        updated = update_variant(qs, self.variant, self.location)
+        self.assertFalse(updated)
+
+        # set a new chip_name
+        self.variant.chip_name = ["test"]
+        updated = update_variant(qs, self.variant, self.location)
+        self.assertTrue(updated)
+
+    def test_update_variant_sequence(self):
+        qs = VariantSheep.objects.filter(
+            name="250506CS3900065000002_1238.1")
+        updated = update_variant(qs, self.variant, self.location)
+        self.assertFalse(updated)
+
+        # update sequence
+        self.variant.sequence = {"test": "AGCT"}
+        updated = update_variant(qs, self.variant, self.location)
+        self.assertTrue(updated)
+
+    def test_update_variant_affymetrix_record(self):
+        qs = VariantSheep.objects.filter(
+            name="250506CS3900065000002_1238.1")
+        updated = update_variant(qs, self.variant, self.location)
+        self.assertFalse(updated)
+
+        # add affymetrix attribute
+        self.variant.probeset_id = "test"
+        self.variant.affy_snp_id = "test"
+        updated = update_variant(qs, self.variant, self.location)
+        self.assertTrue(updated)
+
+    def test_update_variant_with_location(self):
+        qs = VariantSheep.objects.filter(
+            name="250506CS3900065000002_1238.1")
+        updated = update_variant(qs, self.variant, self.location)
+        self.assertFalse(updated)
+
+        # change location
+        self.location.version = "test"
+        updated = update_variant(qs, self.variant, self.location)
+        self.assertTrue(updated)
+
+    def test_update_variant_with_rs_id(self):
+        qs = VariantSheep.objects.filter(
+            name="250506CS3900065000002_1238.1")
+        updated = update_variant(qs, self.variant, self.location)
+        self.assertFalse(updated)
+
+        # update rs_id
+        self.variant.rs_id = "test"
+        updated = update_variant(qs, self.variant, self.location)
+        self.assertTrue(updated)
 
 
 if __name__ == '__main__':